fix: avif heif checked mul (#174)
* fix: check avif/heif decode size overflow

* fix: neon vld1q_u16 oob and chroma gate alignment

* fix: validate tiff strip/tile offsets against mmap

* fix: reject zero raw developed dimensions early

* fix: reject oversized decode dimensions before allocation (DoS guard)

Apply validate_static_decode_dimensions (256 MP cap) before allocating
RGBA buffers in JPEG and RAW decode paths, preventing OOM from
malformed files with absurdly large declared dimensions.

- JPEG baseline: parse SOF header first, validate, then decode
- unpack_libraw_rgb16_rows_to_rgba_f32
- extract_raw_gpu_source
- develop (SDR)
- develop_scene_linear_hdr_with_qual (HDR/f32)
- unpack_thumb (bitmap path)

* fix: add checked_mul and size caps to JXL decode paths (DoS guard)

Validate dimensions via validate_static_decode_dimensions in basic_info,
check buffer sizes against expected values with checked_mul before
allocation, and limit jhgm box to 32 MB.

- JXL_BASIC_INFO: validate_static_decode_dimensions before thread pool setup
- NEED_PREVIEW_OUT_BUFFER: validate preview buffer size vs preview dims
- NEED_IMAGE_OUT_BUFFER: validate main image + K channel buffer sizes
- FULL_IMAGE: expected_len uses checked_mul
- BOX/jhgm: hard cap at 32 MB (was only > usize::MAX check)
- strip_baseline.rs: same BasicInfo + preview size + expected_len checks

* fix: replace wrapping multiplication with checked_rgba_buffer_len in OpenEXR/Radiance allocs (DoS guard)

8 unchecked allocation sites (width * height * 4 without checked_mul)
converted to use crate::constants::checked_rgba_buffer_len:

- openexr_core/read_context.rs: 5 sites (scanline tile, preview,
  mip level, mip preview, chunk RGBA sample count)
- radiance_tiled/header.rs: 1 site (full HDR buffer)
- radiance_tiled/tile_decode.rs: 2 sites (SDR + HDR preview)

* fix: add validate_static_decode_dimensions to AVIF/HEIF/WIC decode paths

Unified 256 MP pixel cap applied at dimension-read entry points to
reject oversized images that pass overflow-only checks.

- avif/decode.rs: avif_checked_pixel_count + lane_count call
  validate_static_decode_dimensions first (covers all AVIF paths)
- heif/decode.rs: 3 functions add cap after zero-size check
  (rgba8_from_planar_rgb444, hdr_buffer_from_interleaved_rgb8_packed,
  hdr_buffer_from_planar_rgb444)
- wic/load.rs: cap after logical_width/logical_height derivation,
  applies to both tiled and full-buffer paths

* fix: replace unchecked multiplications with checked_rgba_buffer_len across compose/gain-map paths and raw_processor

- loader/types.rs: assert_decoded_rgba_len panics on overflow instead of
  silently skipping the length check
- raw_processor.rs: develop() expected_min and rgba_len use checked_mul
- ultra_hdr_compose.rs, jpeg_gain_map_gpu.rs, iso_gain_map_compose_simd.rs,
  jxl_gain_map_deferred.rs, heif/gain_map.rs: replace unchecked
  width*height*4 with checked_rgba_buffer_len (Result/panic/None)

* fix: add frame dimension validation for animation paths and fix PSD test multiplication style

- raster.rs: image_frame_to_static_image_data now returns Result and
  validates frame dimensions before creating DecodedImage
- raster.rs: multi-frame animation path validates each frame's dimensions
  with collect::<Result<Vec<_>, _>>() pattern
- animation_bootstrap.rs: 3 call sites updated for new Result return
- PSD tests: replace (width*height*3) as usize with
  (width as usize)*(height as usize)*3 to avoid wrapping style

* fix: replace panic! in assert_decoded_rgba_len with log::warn! on overflow

The constructor DecodedImage::new has ~100 callers and is not fallible;
panic! on overflow is too harsh. Revert to if-let pattern with a
log::warn! so the length check is silently skipped on overflow
instead of crashing. The actual dimension validation happens at
entry points via validate_static_decode_dimensions.

* fix: use checked_mul for WIC frame selection pixel count

* fix: add checked_mul and validate_static_decode_dimensions to HEIF/JXL/EXR/Radiance/RAW decode paths

HEIF: add validate_static_decode_dimensions to 7 functions missing unified
256MP cap (monochrome, interleaved, YCbCr, etc.)

JXL: fix wrapping multiplication at SUCCESS event; add validate_static to
preview paths in decode.rs and strip_baseline.rs

OpenEXR: fix 6 unchecked w*h*4 multiplications in chromaticities.rs
(deep/IMF flatten, tile extraction), exr_tiled.rs (Vec::with_capacity),
and channels.rs (tile/preview length checks)

Radiance HDR: fix 19 unchecked multiplications across header.rs and
tile_decode.rs — buffer allocations, row offsets, pixel offsets

RAW: add checked_mul + validate_static to load.rs area calculation;
fix 3 unchecked multiplications in raw_processor.rs (Vec allocation,
RGB16 output buffer, bitmap thumb count)

* fix: add validate_static_decode_dimensions to HEIF gain_map decode path

decode_heif_gain_map was using only checked_rgba_buffer_len (overflow
protection) without the unified 256MP cap. Add validate_static_decode_dimensions
after zero-size check. Also fix unchecked row_bytes multiplication.

* fix: add checked_mul to EXR tile copy loop index calculations

extract_rgba32f_tile_from_flat_buffer (chromaticities.rs): src_start,
dst_start, and w4 pre-compute now use checked_mul.

copy_decoded_chunk_to_tile (channels.rs): pre-compute cw4/tw4/cw_copy
strides before loop; src_start/dst_start use checked_mul.

sample_decoded_scanline_chunk_into_preview (channels.rs): pre-compute
src_w4/dst_w4 strides; source_offset and dest_offset use checked_mul.

* fix: add checked mul and dimension validation for image decode safety

* style: cargo fmt

* fix: replace panic and hardcoded values with errors

* refactor: extract duplicate heif_checked_rgba_* functions to mod.rs

- Move heif_checked_rgba_buffer_len / heif_checked_rgba_row_len
  from decode.rs and ycbcr.rs into mod.rs as pub(super) (Rule 14)
- Use super:: prefix in both submodules instead of private copies
- Unify NEON 4:2:0 chroma load annotation unit to 'chroma samples'
- Document validate_static_decode_dimensions return value semantics

* fix: harden decode paths with checked arithmetic

* fix: add checked_mul to ultra_hdr preview paths

+++ src/loader/decode/tests/formats_misc.rs
@@ -41,7 +41,7 @@ fn craft_rgb8_raw_psb(width: u32, height: u32, planar: &[u8]) -> Vec<u8> {
 
 fn craft_rgb8_raw_document(version: u16, width: u32, height: u32, planar: &[u8]) -> Vec<u8> {
     assert!(version == 1 || version == 2);
-    assert_eq!(planar.len(), (width * height * 3) as usize);
+    assert_eq!(planar.len(), (width as usize) * (height as usize) * 3);
     let mut bytes = Vec::new();
     bytes.extend_from_slice(b"8BPS");
     bytes.extend_from_slice(&version.to_be_bytes());
